фикс в команде /changedata
